+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - New `edit` option to customize edit permissions for collections.
 - New `edit` option to customize edit permissions for documents.
 - Permissions options form uploads (`create`, `rename`, `edit`, and `delete`).
+- User-level permissions
 
 ### Removed
 - Float button to edit the current page in the preview.